+++ hass_apps/schedy/__init__.py
@@ -3,4 +3,4 @@ This is the schedy package. Functionality is separated in the
 various sub-modules.
 """
 
-__version__ = "0.8.1"
+__version__ = "0.8.2"